[UPDATE] now working !
- need to fix 127B issue
- need to fix 1G zero issue

+++ md5/includes/ft_ssl_md5.h
@@ -24,7 +24,7 @@
 
 struct s_list;
 typedef struct s_env	t_env;
-typedef uint64_t		t_siz;
+typedef unsigned int		t_siz;
 typedef uint32_t		t_sub;
 
 /*
